Refactoriser la classe LiveResult Fixes #236
